// src/app/admin/interfaces/gindar-info.interface.ts
export interface GindarInfoInterface {
  ruc: string;
  businessReason: string;
  phone1: string;
  phone2: string;
  phone3: string;
  email: string;
  address1: string;
  address2: string;
  descriptionMeta: string;
  facebookLink1: string;
  facebookLink2: string;
  instagramLink: string;
  youtubeLink: string;
  twitterLink: string;
  messengerLink: string;
}